통계 퍼센트 소수점 처리

+++ HamoyeProject/src/main/java/com/bnk/hamoye/domain/Status.java
@@ -7,5 +7,5 @@ public class Status {
 	private String challengeId;
 	private String challengeName;
 	private int userCnt;
-	private int successPercent;
+	private double successPercent;
 }
